+++ containers/Result.h
@@ -52,11 +52,6 @@ namespace evo{
 			};
 
 
-			EVO_NODISCARD constexpr auto value() const&& noexcept -> const T&& {
-				EVO_DEBUG_ASSERT(this->isSuccess());
-				return std::move(*this->val);
-			};
-
 			EVO_NODISCARD constexpr auto value() && noexcept -> T&& {
 				EVO_DEBUG_ASSERT(this->isSuccess());
 				return std::move(*this->val);