+++ containers/Variant.h
@@ -11,14 +11,19 @@ namespace evo{
 	class Variant : public std::variant<Ts...> {
 		public:
 			using std::variant<Ts...>::variant;
+
+			EVO_NODISCARD constexpr auto getNative() const& noexcept -> const std::variant<Ts...>& {
+				return static_cast<const std::variant<Ts...>&>(*this);
+			};
 			
-			EVO_NODISCARD constexpr auto getNative() noexcept -> std::variant<Ts...>& {
+			EVO_NODISCARD constexpr auto getNative() & noexcept -> std::variant<Ts...>& {
 				return static_cast<std::variant<Ts...>&>(*this);
 			};
 
-			EVO_NODISCARD constexpr auto getNative() const noexcept -> const std::variant<Ts...>& {
-				return static_cast<const std::variant<Ts...>&>(*this);
+			EVO_NODISCARD constexpr auto getNative() && noexcept -> std::variant<Ts...>&& {
+				return std::move(static_cast<std::variant<Ts...>&>(*this));
 			};
+
 
 
 			template<class T>
@@ -28,18 +33,26 @@ namespace evo{
 
 
 
+
 			template<class T>
-			EVO_NODISCARD constexpr auto as() noexcept -> T& {
+			EVO_NODISCARD constexpr auto as() const& noexcept -> const T& {
+				debugAssert(this->is<T>(), "Variant is not of this type");
+				return std::get<T>(this->getNative());
+			};
+
+			template<class T>
+			EVO_NODISCARD constexpr auto as() & noexcept -> T& {
 				debugAssert(this->is<T>(), "Variant is not of this type");
 				return std::get<T>(this->getNative());
 			};
 
 
 			template<class T>
-			EVO_NODISCARD constexpr auto as() const noexcept -> const T& {
+			EVO_NODISCARD constexpr auto as() && noexcept -> T&& {
 				debugAssert(this->is<T>(), "Variant is not of this type");
-				return std::get<T>(this->getNative());
+				return std::move(std::get<T>(this->getNative()));
 			};
+
 
 
 			constexpr auto visit(auto callable) noexcept {